Rename package _types -> custom_types and move enum to modules
